rename TitleScreen.tsx to DebugScreen.tsx to reflect new status, add Placeholder Penguins image
